% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idCond_check.r
\name{idcond_check}
\alias{idcond_check}
\alias{idCond_check}
\title{Check the identifiability conditions}
\usage{
idCond_check(P)
}
\arguments{
\item{P}{the source composition matrix in multivariate receptor model}
}
\description{
Check the identifiability conditions C1-C2 of Park and Oh (2015).
}
