% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bmrm.R
\name{plot.bmrm}
\alias{plot.bmrm}
\alias{plot}
\title{Produce plots of the parameter estimates}
\usage{
\method{plot}{bmrm}(x, type = "both", text = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{bmrm}, the output of the function \code{bmrm}}

\item{type}{name of a variable (default="P"). 
It should be one of "P"(source composition or profile matrix P), 
"A"(source contribution matrix A), "both" (both P and A), "Sigma" (error  varaince).}

\item{text}{TRUE/FALSE, display the value of P.hat on the plot for P (defaut=FALSE)}

\item{...}{arguments to be passed to methods}
}
\description{
Produce plots of the estimated 
posterior quantiles of variables for prob=(2.5\%, 50\%, 97.5\%),
based on the MCMC samples in \code{bmrm}.
}
\details{
The following types of plots are drawn depending on the selected parameters:
\itemize{
\item P: bar plots of (5\%, 50\%, 95\%) posterior quantiles of elements for each row of P
\item A: time series plots of (5\%, 50\%, 95\%) posterior quantiles of elements for each column of A
\item Sigma: bar plots of (5\%, 50\%, 95\%) posterior quantiles of of elements of Sigma
 }
}
