% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bmrm.R
\name{summary.bmrm}
\alias{summary.bmrm}
\alias{summary}
\title{Summarize the output of the \code{bmrm} function}
\usage{
\method{summary}{bmrm}(object, digits = 3, ...)
}
\arguments{
\item{object}{an object of class \code{bmrm}, the output of the \code{bmrm} function}

\item{digits}{integer indicating the humber of signifiant digits}

\item{...}{arguments to be passed to methods}
}
\description{
An S3 method that summarizes the
output of the \code{bmrm} function in an object of class \code{bmrm}. This object
 contains  
the posterior mean, the posterior standard deviation, and 
(0.025, 0.05, 0.1,0.25, 0.5, 0.75,0.9, 0.95, 0.975) posterior quantiles
 of  A, P, \eqn{\Sigma}. It also contains other
relevant information about the MCMC procedure such as the burn-in iterations, 
the number of MCMC chains, etc.
}
