% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_ACF_plot.R
\name{trace_ACF_plot}
\alias{trace_ACF_plot}
\title{Trace and/or ACF plots of elements of a variable
 in \code{bmrm} object}
\usage{
trace_ACF_plot(x,var="P",ACF=FALSE, nplot=12,saveFile=FALSE,...)
}
\arguments{
\item{x}{an object of class \code{bmrm}, the output of the \code{bmrm} function}

\item{var}{name of a variable to which the plots apply.  It sould be one of
"A" (source contribution matrix),
"P" (source composition or profile matrix),
"Sigma" (error variance).}

\item{ACF}{TRUE/FALSE  IF TRUE ACF plot will be provided along with trace
plot (dafault: FALSE)}

\item{nplot}{number of elements of 'var' for trace/ACF plot. If 'nplot' is
smaller than the total number of elements of 'var' then plots of the first
'nplot' elements will be drawn. Otherwise, trace/ACF plots of  all
 elements will be drawn.
  (default=0  implies that all elements will be selected if var="P" or "Sigma"
  and the first 12 elements will be selected  if  var="A")}

\item{saveFile}{TRUE/FALSE, save the plots in file
\emph{'var'-trace.pdf} (default=FALSE)}

\item{...}{arguments to be passed to methods}
}
\description{
Produce trace and Auto-Correlation Function
  plots of MCMC samples
 of  elements of A, P, \eqn{Sigma}.
}
\examples{
\dontrun{
data(Elpaso); Y=Elpaso$Y ; muP=Elpaso$muP ; q=nrow(muP)
out.Elpaso <- bmrm(Y,q,muP, nAdapt=1000,nBurnIn=5000,nIter=5000,nThin=1)
trace_ACF_plot(out.Elpaso,"Sigma")
trace_ACF_plot(out.Elpaso,"P", ACF=T, saveFile=TRUE)
trace_ACF_plot(out.Elpaso,"A", nplot=12)
}

}
