% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_ACF__plot_indiv.R
\name{trace_ACF_plot_indiv}
\alias{trace_ACF_plot_indiv}
\title{Trace and ACF plots of an element of a variable in \code{bmrm} object}
\usage{
trace_ACF_plot_indiv(x,var="P",sourceID=1,varID=NULL,obsID=1,...)
}
\arguments{
\item{x}{an object of class \code{bmrm}, the output of the \code{bmrm} function}

\item{var}{name of a variable. It sould be one of
"A" (source contribution matrix),
"P" (source composition or profile matrix),
"Sigma" (error variance).}

\item{sourceID}{source index (default=1)}

\item{varID}{variable index (default=1)}

\item{obsID}{observation index (default=1)}

\item{...}{arguments to be passed to methods}
}
\description{
Produce trace and Auto-Correlation Function plots of MCMC samples
 of a specific element of A, P, \eqn{\Sigma}.
}
\details{
Produce trace and Auto-Correlation Function plots of MCMC samples
 of a specific element of A or P or \eqn{\Sigma}. The element is specified by
 the indices
 \code{sourceID}, \code{varID}, and \code{obsID} of variable \code{var}.
}
\examples{
\dontrun{
data(Elpaso)
Y=Elpaso$Y ; muP=Elpaso$muP
q=nrow(muP)
out.Elpaso <- bmrm(Y,q,muP, nAdapt=1000,nBurnIn=5000,nIter=5000,nThin=1)
trace_ACF_plot_indiv(out.Elpaso,"P")
trace_ACF_plot_indiv(out.Elpaso,"P",sourceID=2, varID=3 )
trace_ACF_plot_indiv(out.Elpaso,"A", obsID=10, sourceID=2 )
trace_ACF_plot_indiv(out.Elpaso,"Sigma", varID=3)
}

}
