% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsACF_indiv_plot_bmrm.R
\name{tsACF_indiv_plot_bmrm}
\alias{tsACF_indiv_plot_bmrm}
\title{Time series and ACF plots of an element of a variable in \code{bmrm} object}
\usage{
tsACF_indiv_plot_bmrm(x,var="P",sourceID=1,varID=NULL,obsID=1,...)
}
\arguments{
\item{x}{an object of class \code{bmrm}, the output of the \code{bmrm} function}

\item{var}{name of a variable. It sould be one of 
"A" (source contribution matrix), 
"P" (source composition or profile matrix), 
"Sigma" (error variance).}

\item{sourceID}{source index (default=1)}

\item{varID}{variable index (default=1)}

\item{obsID}{observation index (default=1)}

\item{...}{arguments to be passed to methods}
}
\description{
Produce time series and Auto-Correlation Function plots of MCMC samples
 of a specific element of A or  P or \eqn{\Sigma}. The effective sample size 
 proposed by R. Neal (Kass et al. 1998, p.99) is also provided.
}
\details{
Produce time series and Auto-Correlation Function plots of MCMC samples
 of a specific element of A or P or \eqn{\Sigma}. The element is specified by
 the indices   
 \code{sourceID}, \code{varID}, and \code{obsID} of variable \code{var}. 
 @references Kass, R.E., Carlin, B.P., Gelman, R., Neal, R.M. (1998), 
 Markov chain Monte Carlo in practice: A roundtable discussion, The American
 Statistician, 52, 93-100.
}
\examples{
\dontrun{
data(simdata)
T = nrow(simdata); J = ncol(simdata)
q = 3
muP=matrix(0.5, K,J)
muP[1,c(2,3)]<-muP[2,c(1,4)]<-muP[3,c(5,6)]<-0 
muP<-t(apply(muP,1,function(x) x/sum(x)))
out <- bmrm(simdata,q, muP,nAdapt=1000,nBurnIn=1000,nIter=3000, nThin=1)
tsACF_indiv_plot_bmrm(out,"P")
tsACF_indiv_plot_bmrm(out,"P",souceID=2, varID=3 )
tsACF_indiv_plot_bmrm(out,"A", obsID=10, souceID=2 )
#' tsACF_indiv_plot_bmrm(out,"Sigma", varID=3)
}

}
