% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsACF_plot_bmrm.R
\name{tsACF_plot_bmrm}
\alias{tsACF_plot_bmrm}
\title{Time series and ACF plots of randomly selected elements of a variable
 in \code{bmrm} object}
\usage{
tsACF_plot_bmrm(x,var="P",nplot=3,saveFile=FALSE,...)
}
\arguments{
\item{x}{an object of class \code{bmrm}, the output of the \code{bmrm} function}

\item{var}{name of a variable to which the plots apply.  It sould be one of 
"A" (source contribution matrix), 
"P" (source composition or profile matrix), 
"Sigma" (error variance).}

\item{nplot}{number of elements to be selected randomly (default=3)}

\item{saveFile}{TRUE/FALSE, save the plots in file 
\emph{'var'-trace.pdf} (default=FALSE)}

\item{...}{arguments to be passed to methods}
}
\description{
Produce time series and Auto-Correlation Function
  plots of MCMC samples 
 of randomly selected element of A or P or Sigma. The effective sample size 
 proposed by R. Neal (Kass et al. 1998, p.99) is also provided.
}
\details{
produces time series and auto-correlation function plots of
the MCMC samples of randomly selected \code{'nplot'} elements of \code{'var'} 
in \code{bmrm} object \code{'x'}. 
 @references Kass, R.E., Carlin, B.P., Gelman, R., Neal, R.M. (1998), 
 Markov chain Monte Carlo in practice: A roundtable discussion, The American
 Statistician, 52, 93-100.
}
\examples{
\dontrun{
data(simdata)
T = nrow(simdata); J = ncol(simdata)
q = 3
muP=matrix(0.5, K,J)
muP[1,c(2,3)]<-muP[2,c(1,4)]<-muP[3,c(5,6)]<-0 
muP<-t(apply(muP,1,function(x) x/sum(x)))
out <- bmrm(simdata,q, muP,nAdapt=1000,nBurnIn=1000,nIter=3000, nThin=1)
tsACF_plot_bmrm(out,"Sigma")
tsACF_plot_bmrm(out,"P",saveFile=TRUE)
tsACF_plot_bmrm(out,"A",nplot=2)
}

}
