% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convdiag_bmrm.R
\name{convdiag_bmrm}
\alias{convdiag_bmrm}
\title{Convergence Diagnostics on MCMC samples in \code{bmrm}}
\usage{
convdiag_bmrm(x , var="P", convdiag="heidel",print=TRUE,...)
}
\arguments{
\item{x}{an object of class \code{bmrm}, the output of the \code{bmrm} function}

\item{var}{name of a variable for convergence disagnostics. It should be one of "A" (source contribution matrix), 
"P" (source composition or profile matrix), or
"Sigma" (error variance).}

\item{convdiag}{vector of convergence diagnostic measures. It should be any subvector
of ("geweke", "heidel") (default="heidel").}

\item{print}{TRUE/FALSE, print convergence diagnostics results (default=TRUE)}

\item{...}{arguments to be passed to methods}
}
\value{
A list of  convergence diagnostics results
\describe{
  \item{convdiag}{selected convergence diagnostic measures} 
  \item{geweke}{Geweke's statistics and p-value if \code{convdiag}
 includes "geweke", NULL if \code{convdiag} does not include "geweke"} 
  \item{heidel}{Heidelberger and Welch's test statistics 
and p-value if \code{convdiag} includes "heidel"; NULL if 
\code{convdiag} does not include "heidel"} 
}
}
\description{
Compute convergence diagnostic measures of 
 Geweke (1992), Heidelberger and Welch (1983).
}
\details{
The functions  \code{geweke.diag}, \code{heidel.diag} 
 in \bold{coda} package are used to compute these diagnostics. 
 In our package, \code{frac1=0.1, frac2=0.5} are used as parameters of the function
 \code{coda::geweke.diag} and \code{eps = 0.1, pvalue = 0.05} are used as
  parameters of the function \code{coda::heidel.diag}.
}
\examples{
\dontrun{
data(simdata)
T = nrow(simdata); J = ncol(simdata)
q = 3
muP=matrix(0.5, q,J)
muP[1,c(2,3)]<-muP[2,c(1,4)]<-muP[3,c(5,6)]<-0 
muP<-t(apply(muP,1,function(x) x/sum(x)))
out <- bmrm(simdata,q,muP,nAdapt=1000,nBurnIn=1000,nIter=3000,nThin=1)

conv1<-convdiag_bmrm(out,var="P")
conv2<-convdiag_bmrm(out,var="A", convdiag="geweke")
conv3<-convdiag_bmrm(out,var="Sigma", convdiag=c("geweke","heidel"))
}
}
\references{
Geweke, J.(1992) Evaluating the accuracy of sampling-based
approaches to calculating posterior moments. In Bayesian Statistics 4
(ed JM Bernado, JO Berger, AP Dawid and AFM Smith). Clarendon Press.

Heidelberger P, and Welch PD. (1981) A spectral method for
confidence interval generation and run length control in simulations.
Comm. ACM. 24, 233-245.

Heidelberger P. and Welch PD.(1983) Simulation run length
control in the presence of an initial transient.
Opns Res., 31, 1109-44,Oxford, UK.

Plummer, M., Best, N., Cowles, K. and Vines K. (2006) CODA: 
Convergence Diagnosis and Output Analysis for MCMC, R News, Vol 6, pp. 7-11.
}
